nans are still hard but they work (?) now

+++ magicbean/pocket.go
@@ -30,7 +30,7 @@ type Pocket struct {
 	bottles, stones, meds, rewards, notes []zecs.Entity
 }
 
-func (this Pocket) Has(entity zecs.Entity, n uint64) bool {
+func (this Pocket) Has(entity zecs.Entity, n float64) bool {
 	return this.inventory.Count(entity) >= n
 }
 
@@ -56,19 +56,19 @@ func (this Pocket) HasBottle() bool {
 	return this.HasAny(this.bottles)
 }
 
-func (this Pocket) HasStones(n uint64) bool {
+func (this Pocket) HasStones(n float64) bool {
 	return this.inventory.Sum(this.stones) >= n
 }
 
-func (this Pocket) HasMedallions(n uint64) bool {
+func (this Pocket) HasMedallions(n float64) bool {
 	return this.inventory.Sum(this.meds) >= n
 }
 
-func (this Pocket) HasDungeonRewards(n uint64) bool {
+func (this Pocket) HasDungeonRewards(n float64) bool {
 	return this.inventory.Sum(this.rewards) >= n
 }
 
-func (this Pocket) HasHearts(n uint64) bool {
+func (this Pocket) HasHearts(n float64) bool {
 	pieces := this.inventory.Count(this.heartPiece)
 	hearts := pieces / 4
 	return hearts >= n
